+++ noodles-util/src/alignment.rs
@@ -3,8 +3,9 @@
 #[cfg(feature = "async")]
 pub mod r#async;
 
+mod index;
 pub mod io;
 pub mod iter;
 mod record;
 
-pub use self::record::Record;
+pub use self::{index::Index, record::Record};